Factor out identifier-type handling
